bary2cart {geometry}	R Documentation

Conversion of Barycentric to Cartesian coordinates

Description

Given the barycentric coordinates of one or more points with respect to a simplex, compute the Cartesian coordinates of these points.

Usage

bary2cart(X, Beta)

Arguments

X
Reference simplex in N dimensions represented by a N+1-by-N matrix

Beta
M points in barycentric coordinates with respect to the simplex X represented by a M-by-N+1 matrix

Value

M-by-N matrix in which each row is the Cartesian coordinates of corresponding row of Beta

Author(s)

David Sterratt

See Also

cart2bary

Examples

## Define simplex in 2D (i.e. a triangle)
X <- rbind(c(0, 0),
           c(0, 1),
           c(1, 0))
## Cartesian cooridinates of points
beta <- rbind(c(0, 0.5, 0.5),
              c(0.1, 0.8, 0.1))
## Plot triangle and points
trimesh(rbind(1:3), X)
text(X[,1], X[,2], 1:3) # Label vertices
P <- bary2cart(X, beta)
points(P)

[Package geometry version 0.4.5 Index]