cart2pol {geometry}	R Documentation

Transform Cartesian coordinates to polar or cylindrical coordinates.

Description

The inputs x, y (, and z) must be the same shape, or scalar. If called with a single matrix argument then each row of C represents the Cartesian coordinate (x, y (, z)).

Usage

cart2pol(x, y = NULL, z = NULL)

Arguments

x
x-coordinates or matrix with three columns

y
y-coordinates (optional, if x) is a matrix

z
z-coordinates (optional, if x) is a matrix

Value

A matrix P where each row represents one polar/(cylindrical) coordinate (theta, r, (, z)).

Author(s)

Kai Habel

David Sterratt

See Also

pol2cart, cart2sph, sph2cart

[Package geometry version 0.4.1.1 Index]