cart2sph {geometry}	R Documentation

If called with a single matrix argument then each row of c represents the Cartesian coordinate (x, y, z).

Description

Transform Cartesian to spherical coordinates

Usage

cart2sph(x, y = NULL, z = NULL)

Arguments

x
x-coordinates or matrix with three columns

y
y-coordinates (optional, if x) is a matrix

z
z-coordinates (optional, if x) is a matrix

Value

Matrix with columns:

theta
the angle relative to the positive x-axis

phi
the angle relative to the xy-plane

r
the distance to the origin (0, 0, 0)

Author(s)

Kai Habel

David Sterratt

See Also

sph2cart, cart2pol, pol2cart

[Package geometry version 0.4.5 Index]