dot {geometry}	R Documentation

Compute the dot product of two vectors

Description

If x and y are matrices, calculate the dot-product along the first non-singleton dimension. If the optional argument d is given, calculate the dot-product along this dimension.

Usage

dot(x, y, d = NULL)

Arguments

x
Matrix of vectors

y
Matrix of vectors

d
Dimension along which to calculate the dot product

Value

Vector with length of dth dimension

Author(s)

David Sterratt

[Package geometry version 0.4.5 Index]