entry.value {geometry}	R Documentation

Retrieve or set a list of array element values

Description

entry.value retrieves or sets the values in an array a at the positions indicated by the rows of a matrix idx.

Usage

entry.value(a, idx)

Arguments

a
An array.

idx
Numerical matrix with the same number of columns as the number of dimensions of a. Each row indices a cell in a of which the value is to be retrieved or set.

value
An array of length nrow(idx).

Value

entry.value(a,idx) returns a vector of values at the indicated cells. entry.value(a,idx) <- val changes the indicated cells of a to val.

Author(s)

Raoul Grasman

Examples

a = array(1:(4^4),c(4,4,4,4))
entry.value(a,cbind(1:4,1:4,1:4,1:4))
entry.value(a,cbind(1:4,1:4,1:4,1:4)) <- 0

entry.value(a, as.matrix(expand.grid(1:4,1:4,1:4,1:4)))
     # same as `c(a[1:4,1:4,1:4,1:4])' which is same as `c(a)'

[Package geometry version 0.4.1.1 Index]