extprod3d {geometry}	R Documentation

Compute external- or ‘cross’- product of 3D vectors.

Description

Computes the external product

(x2 * y3 - x3 * y2, x3 * y1 - x1 * y3, x1 * y2 - x2 * y1)

(x2 * y3 - x3 * y2, x3 * y1 - x1 * y3, x1 * y2 - x2 * y1)

(x2 * y3 - x3 * y2, x3 * y1 - x1 * y3, x1 * y2 - x2 * y1)

of the 3D vectors in x and y.

Usage

extprod3d(x, y, drop = TRUE)

Arguments

x
n-by-3 matrix. Each row is one x-vector

y
n-by-3 matrix. Each row is one y-vector

drop
logical. If TRUE and if the inputs are one row matrices or vectors, then delete the dimensions of the array returned.

Value

If n is greater than 1 or drop is FALSE, n-by-3 matrix; if n is 1 and drop is TRUE, a vector of length 3.

Author(s)

Raoul Grasman

See Also

drop

[Package geometry version 0.4.1.1 Index]