feasible.point {geometry}	R Documentation

Find point in intersection of convex hulls

Description

Find point that lies somewhere in interesction of two convex hulls. If such a point does not exist, return NA. The feasible point is found using a linear program similar to the one suggested at ../doc/qhull/html/qhalf.html#notes

Usage

feasible.point(ch1, ch2, tol = 0)

Arguments

ch1
First convex hull with normals

ch2
Second convex hull with normals

tol
The point must be at least this far within the facets of both convex hulls

[Package geometry version 0.4.5 Index]