halfspacen {geometry}	R Documentation

Compute halfspace intersection about a point

Description

Compute halfspace intersection about a point

Usage

halfspacen(p, fp, options = "Tv")

Arguments

p
An M-by-N+1 matrix. Each row of p represents a halfspace by a N-dimensional normal to a hyperplane and the offset of the hyperplane.

fp
A “feasible” point that is within the space contained within all the halfspaces.

options
String containing extra options, separated by spaces, for the underlying Qhull command; see Qhull documentation at ../doc/qhull/html/qhalf.html.

Value

A N-column matrix containing the intersection points of the hyperplanes ../doc/qhull/html/qhalf.html.

Note

halfspacen was introduced in geometry 0.4.0, and is still under development. It is worth checking results for unexpected behaviour.

Author(s)

David Sterratt

References

Barber, C.B., Dobkin, D.P., and Huhdanpaa, H.T., “The Quickhull algorithm for convex hulls,” ACM Trans. on Mathematical Software, Dec 1996.

http://www.qhull.org

See Also

convhulln

Examples

p <- rbox(0, C=0.5)  # Generate points on a unit cube centered around the origin
ch <- convhulln(p, "n") # Generate convex hull, including normals to facets, with "n" option
# Intersections of half planes
# These points should be the same as the orginal points
pn <- halfspacen(ch$normals, c(0, 0, 0)) 

[Package geometry version 0.4.1.1 Index]