mesh.dcircle {geometry}	R Documentation

Circle distance function

Description

Signed distance from points p to boundary of circle to allow easy definition of regions in distmesh2d.

Usage

mesh.dcircle(p, radius = 1, ...)

Arguments

p
A matrix with 2 columns (3 in mesh.dsphere), each row representing a point in the plane.

radius
radius of circle

...
additional arguments (not used)

Value

A vector of length nrow(p) containing the signed distances to the circle

Author(s)

Raoul Grasman; translated from original Matlab sources of Per-Olof Persson.

References

http://persson.berkeley.edu/distmesh/

P.-O. Persson, G. Strang, A Simple Mesh Generator in MATLAB. SIAM Review, Volume 46 (2), pp. 329-345, June 2004

See Also

distmesh2d, mesh.drectangle, mesh.diff, mesh.intersect, mesh.union

Examples

example(distmesh2d)

[Package geometry version 0.4.5 Index]