mesh.diff {geometry}	R Documentation

Difference, union and intersection operation on two regions

Description

Compute the signed distances from points p to a region defined by the difference, union or intersection of regions specified by the functions regionA and regionB. regionA and regionB must accept a matrix p with 2 columns as their first argument, and must return a vector of length nrow(p) containing the signed distances of the supplied points in p to their respective regions.

Usage

mesh.diff(p, regionA, regionB, ...)

Arguments

p
A matrix with 2 columns (3 in mesh.dsphere), each row representing a point in the plane.

regionA
vectorized function describing region A in the union / intersection / difference

regionB
vectorized function describing region B in the union / intersection / difference

...
additional arguments passed to regionA and regionB

Value

A vector of length nrow(p) containing the signed distances to the boundary of the region.

Author(s)

Raoul Grasman; translated from original Matlab sources of Per-Olof Persson.

See Also

distmesh2d, mesh.dcircle, mesh.drectangle mesh.dsphere

[Package geometry version 0.4.1.1 Index]