mesh.drectangle {geometry}	R Documentation

Rectangle distance function

Description

Signed distance from points p to boundary of rectangle to allow easy definition of regions in distmesh2d.

Usage

mesh.drectangle(p, x1 = -1/2, y1 = -1/2, x2 = 1/2, y2 = 1/2, ...)

Arguments

p
A matrix with 2 columns, each row representing a point in the plane.

x1
lower left corner of rectangle

y1
lower left corner of rectangle

x2
upper right corner of rectangle

y2
upper right corner of rectangle

...
additional arguments (not used)

Value

a vector of length nrow(p) containing the signed distances

Author(s)

Raoul Grasman; translated from original Matlab sources of Per-Olof Persson.

References

http://persson.berkeley.edu/distmesh/

P.-O. Persson, G. Strang, A Simple Mesh Generator in MATLAB. SIAM Review, Volume 46 (2), pp. 329-345, June 2004

See Also

distmesh2d, mesh.drectangle, mesh.diff, mesh.intersect, mesh.union

Examples

example(distmesh2d)

[Package geometry version 0.4.1.1 Index]