mesh.dsphere {geometry}	R Documentation

Sphere distance function

Description

Signed distance from points p to boundary of sphere to allow easy definition of regions in distmeshnd.

Usage

mesh.dsphere(p, radius = 1, ...)

Arguments

p
A matrix with 2 columns (3 in mesh.dsphere), each row representing a point in the plane.

radius
radius of sphere

...
additional arguments (not used)

Value

A vector of length nrow(p) containing the signed distances to the sphere

Author(s)

Raoul Grasman; translated from original Matlab sources of Per-Olof Persson.

References

http://persson.berkeley.edu/distmesh/

P.-O. Persson, G. Strang, A Simple Mesh Generator in MATLAB. SIAM Review, Volume 46 (2), pp. 329-345, June 2004

See Also

distmeshnd

Examples

example(distmeshnd)

[Package geometry version 0.4.1.1 Index]