pol2cart {geometry}	R Documentation

Transform polar or cylindrical coordinates to Cartesian coordinates.

Description

The inputs theta, r, (and z) must be the same shape, or scalar. If called with a single matrix argument then each row of P represents the polar/(cylindrical) coordinate (theta, r (, z)).

Usage

pol2cart(theta, r = NULL, z = NULL)

Arguments

theta
describes the angle relative to the positive x-axis.

r
is the distance to the z-axis (0, 0, z).

z
(optional) is the z-coordinate

Value

a matrix C where each row represents one Cartesian coordinate (x, y (, z)).

Author(s)

Kai Habel

David Sterratt

See Also

cart2pol, sph2cart, cart2sph

[Package geometry version 0.4.1.1 Index]