polyarea {geometry}	R Documentation

Determines area of a polygon by triangle method.

Description

Determines area of a polygon by triangle method. The variables x and y define the vertex pairs, and must therefore have the same shape. They can be either vectors or arrays. If they are arrays then the columns of x and y are treated separately and an area returned for each.

Usage

polyarea(x, y, d = 1)

Arguments

x
X coordinates of vertices.

y
Y coordinates of vertices.

d
Dimension of array to work along.

Details

If the optional dim argument is given, then polyarea works along this dimension of the arrays x and y.

Value

Area(s) of polygon(s).

Author(s)

David Sterratt based on the octave sources by David M. Doolin

Examples

x <- c(1, 1, 3, 3, 1)
y <- c(1, 3, 3, 1, 1)
polyarea(x, y)
polyarea(cbind(x, x), cbind(y, y)) ##  c(4, 4)
polyarea(cbind(x, x), cbind(y, y), 1) ##  c(4, 4)
polyarea(rbind(x, x), rbind(y, y), 2) ##  c(4, 4)

[Package geometry version 0.4.1.1 Index]