sph2cart {geometry}	R Documentation

Transform spherical coordinates to Cartesian coordinates

Description

The inputs theta, phi, and r must be the same shape, or scalar. If called with a single matrix argument then each row of S represents the spherical coordinate (theta, phi, r).

Usage

sph2cart(theta, phi = NULL, r = NULL)

Arguments

theta
describes the angle relative to the positive x-axis.

phi
is the angle relative to the xy-plane.

r
is the distance to the origin (0, 0, 0).

If only a single return argument is requested then return a matrix C where each row represents one Cartesian coordinate (x, y, z).

Author(s)

Kai Habel

David Sterratt

See Also

cart2sph, pol2cart, cart2pol

[Package geometry version 0.4.5 Index]