trimesh {geometry}	R Documentation

Display triangles mesh (2D)

Description

trimesh(T, p) displays the triangles defined in the m-by-3 matrix T and points p as a mesh. Each row of T specifies a triangle by giving the 3 indices of its points in X.

Usage

trimesh(T, p, p2, add = FALSE, axis = FALSE, boxed = FALSE, ...)

Arguments

T
T is a m-by-3 matrix. A row of T contains indices into X of the vertices of a triangle. T is usually the output of delaunayn.

p
A vector or a matrix.

p2
if p is not a matrix p and p2 are bind to a matrix with cbind.

add
Add to existing plot in current active device?

axis
Draw axes?

boxed
Plot box?

...
Parameters to the rendering device. See the rgl package.

Author(s)

Raoul Grasman

See Also

tetramesh, rgl, delaunayn, convhulln, surf.tri

Examples

#example trimesh
p = cbind(x=rnorm(30), y=rnorm(30))
tt = delaunayn(p)
trimesh(tt,p)

[Package geometry version 0.4.1.1 Index]