mesh.hunif {geometry}	R Documentation

Uniform desired edge length

Description

Uniform desired edge length function of position to allow easy definition of regions when passed as the fh argument of distmesh2d or distmeshnd.

Usage

mesh.hunif(p, ...)

Arguments

p
A n-by-m matrix, each row representing a point in an m-dimensional space.

...
additional arguments (not used)

Value

Vector of ones of length n.

Author(s)

Raoul Grasman; translated from original Matlab sources of Per-Olof Persson.

See Also

distmesh2d and distmeshnd.

[Package geometry version 0.4.5 Index]