tetramesh {geometry}	R Documentation

Render tetrahedron mesh (3D)

Description

tetramesh(T, X, col) uses the rgl package to display the tetrahedrons defined in the m-by-4 matrix T as mesh. Each row of T specifies a tetrahedron by giving the 4 indices of its points in X.

Usage

tetramesh(T, X, col = grDevices::heat.colors(nrow(T)), clear = TRUE,
  ...)

Arguments

T
T is a m-by-3 matrix in trimesh and m-by-4 in tetramesh. A row of T contains indices into X of the vertices of a triangle/tetrahedron. T is usually the output of delaunayn.

X
X is an n-by-2/n-by-3 matrix. The rows of X represent n points in 2D/3D space.

col
The tetrahedron colour. See rgl documentation for details.

clear
Should the current rendering device be cleared?

...
Parameters to the rendering device. See the rgl package.

Author(s)

Raoul Grasman

See Also

trimesh, rgl, delaunayn, convhulln, surf.tri

Examples

## Not run: 
# example delaunayn
d = c(-1,1)
pc = as.matrix(rbind(expand.grid(d,d,d),0))
tc = delaunayn(pc)

# example tetramesh
clr = rep(1,3) %o% (1:nrow(tc)+1)
rgl::rgl.viewpoint(60,fov=20)
rgl::rgl.light(270,60)
tetramesh(tc,pc,alpha=0.7,col=clr)

## End(Not run)

[Package geometry version 0.4.1.1 Index]